requirements.txt
# Core dependencies
biopython>=1.81
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
keras>=2.12.0

# Graph Neural Networks
torch-geometric>=2.3.0
dgl>=1.1.0

# Bioinformatics tools integration
pysam>=0.21.0
khmer>=3.0.0
screed>=1.1.2

# Sequence analysis
jellyfish>=0.11.0
kmc>=3.2.1

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
networkx>=2.8.0

# CLI and configuration
click>=8.0.0
pyyaml>=6.0
toml>=0.10.2
rich>=12.0.0
tqdm>=4.64.0

# Database and web access
requests>=2.28.0
urllib3>=1.26.0
entrez-direct>=16.2

# Parallel processing
joblib>=1.1.0
multiprocessing-logging>=0.3.4

# Quality control
fastqc>=0.11.9
trimmomatic>=0.39

# Assembly tools (external dependencies - install separately)
# spades>=3.15.0
# novoplasty>=4.3.1
# getorganelle>=1.7.7

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991